Test: print note when we're building container
Building the container the first time can take a good 15 seconds or more,
which amount of silence is likely to be worrisome to someone running the
Test script. So we print a line indicating we've started the build and that
it may take a few moments. This adds an extra line to the output, but
that's hardly noticable even when building and testing as quietly as
possible, which still produces a twenty or more lines of output.