Test: Add -q option to run all tests as quietly as possible
This is an option because it can make it very difficult to understand
failures, so we only want quiet mode used those who know to run it again
without the quiet flag if something goes wrong in a non-obvious way.

The Docker image build is also now more verbose when -q isn't specified.
but we leave the "Building docker image" message in place anyway because
it's still mildly helpful with -q and increases output by only about 5%.