Work with Python versions <3.6; print version info on bootstrap install
For Python versions < 3.6 we need to use a slightly different get-pip.py,
so we check the version and include it in the URL if it's < 3.6. This works
with 2.7 and 3.5; we've not checked to see if there are other versions
available.

The only way to test this at the moment is to change the `python3` to
`python2` in the pactivate code, but it will get an automatic test when
we tweak pactivate to allow Python version selection by the user.