pactivate: separate bootstrap dirs for non-modern (<3.6) Python versions
We were already downloading different pip bootstraps for Python ≤3.5
because the current version of get_pip.py has dropped compatibility
with those earlier versions of Python. To avoid any chance of using
the wrong bootstrap code we now name the bootstrap directory for the
edition of get_pip.py we downloaded (Python major and minor version
number). This also may speed things by re-using a bootstrap when the
developer switches Python versions.

This change was considerably more difficult to implement than
expected, as the amount of changed code will indicate. The main
problem was that since we now need to get the Python version _before_
we determine the bootstrap directory name we need to quiet down (when
`-q` is given) all the verbosity of that process that previously
happened only once during bootstrap, rather than every time.

Other changes:
- We move the test of -q to an earlier test so we don't need to use it
  in the "run pactivate with options" test; this makes debugging the
  latter test easier.
- This includes some cleanup of cont-test to note the start and end of
  the test run and better name variables.

This really should be warning when the current virtualenv is not using
the same Python as the one the script would use for a fresh install of
that virtualenv; that's added to the to-do list.