pactivate: Improve messages about what Python version is being used
We determine the Python version before we look for a bootstrap directory,
but we know that this is the Python version we'll be using _only_ if we
later generate the virtualenv. If a virtualenv already exists, we'll be
using whatever Python version is already set up in that, regardless of what
$BASE/.python or python3 runs.

So to avoid confusion, we no longer display the version we _would_ use to
create a virtualenv, but instead display that version information only if
we _are actually creating_ a virtualenv. And since this now rarely appears,
we have no need to suppress these in quiet mode, which makes quiet mode a
bit more informational while still suppressing a large amount of verbosity.

We also now display the symlink information in a slightly nicer way.