pactivate v0.2: Don't break virtualenv if $BASE/.python is changed
For Python 3 the virtualenv package makes virtualenv/bin/python a symlink
to the real interpreter (at least on Unix), rather than copying the file as
Python 2 does. Thus, if you used a $BASE/.python symlink pointing to the
interpreter, removing that symlink breaks the virtualenv in an obvious way
(python won't run) and changing it to point to a different interpreter may
break the virtualenv in subtle ways.

So now if $BASE/.python is a symlink we dereference it and use the absolute
path of its target as our Python interpreter, removing any dependency on
$BASE/.python after the virtualenv is created. (We still print a warning,
as usual, if .python and virtualenv/bin/python are different.)