pactivate: Fix version number check for "latest" or versioned pip
The latest version of pip does not work on Python versions ≤ 3.6, so
we must check the version of Python we're using to determine if we use
"latest" or "3.5" or whatever.

Unfortunately, our check did a string comparison of
platform.python_version() which broke when 3.10 was released because
"3.10" < "3.6". Fix this by doing a proper integer compare of major
and minor versions using the values from sys.version_info.

Note that we can't use `M, m, *_ = sys.version_info` or similar syntax
because we must be compatible with Python 2.