WIP noncont-test: Extract tests that don't need a container
It's slow to bring up a container in which to run the tests so we
extract those tests that don't need a container (i.e., don't depend
on a specific Python version or location) to a separate file that
is run without a container and can be run independently. Thus we can
do a quick test of part of the system by running `./noncont-test`.

We move the stored output of the tests into a subdirectory (the $to
variable) to avoid polluting the top-level directory. This wasn't an
issue before because the output files were in a container rather than
the project directory. We also now properly quote uses of $BUILD and $to
because we may now have a space in the (user-chosen) path leading to
these.

TODO:
- ???