Test,cont-test: Move argument parsing into cont-test
We're planning to add a second "non-cont-test" script that will let us
move some tests out of the container, and we want to be able to run that
one "by hand" without running all of the container tests as well. Thus,
the scripts called by Test should take the same command-line parameters
that Test itself does.

The environment variables set up by Test now can disappear, but Test
still needs the '-q' argument for `docker build`, but should ignore
any other arguments (letting the sub-scripts show a failure if any
arguments are bad). We use a little trick from [1] to see if there's
a '-q' among any of the arguments passed in.

[1]: https://stackoverflow.com/a/22658129/107294

Reviewed-by: Yasuaki Kudo <yasu@yasuaki.com>