cont-test: Fix wrong message and wrong filename for a Bash test
The "error when run with /bin/bash" test was not checking for the
correct error message. Unfortunately, it was also checking against
the wrong file containing the message (the file from the previous test)
which had the incorrect message in it. So the code was ok, but the test
was not actually testing it.