README: Rename "base" to "project" directory
This makes it a little more clear what sort of "base" directory we're
talking about.

Potentially we should consider changing the `-B` option to `-P` or
something similar, to better match this. However, we want to avoid
confusion with `-p`, which is often used to specify the path to a specific
Python interpreter to use in the virtualenv. (We no longer support a `-p`
option ourselves, as this functionality is already provided by the
`$PROJDIR/.python` symlink, but other programs do use `-p` for the same
purpose.)