pactivate 0.3: Python 3.6.x no longer uses latest get-pip.py
The latest version of get-pip.py no longer works with Python 3.6.x.
(This may be because 3.6 end-of-life was 2021-12-23, about nine
months ago.) So our version check is now tweaked to use the
get-pip.py specifically for 3.6, rather than the generic version.

This was tested by hand with Pythons 3.6.15, 3.7.14 and 3.5.10; we
currently have no automated tests for versions of Python other than
2.7 and 3.9. In the development branch we have started work to be able
to more easily test these old 3.x versions of Python.