pactivate: No longer overwrite $ve and $pa shell vars in some circumstances
pactivate checks to see if $BUILD/virtualenv/bin/python and .python
are the same file and prints a warning if they're not. We constructed
the warning using two variables, $ve and $pa, which overwrote any
previous values they had in the calling shell (or left them set if
they were not previously set).

Avoid this by constructing the error message without using variables.