WIP pactivate: Fix and improve bin/ vs. Script/ handling in virtualenv
The bootstrap limited virtual environment that we create with `pip
install -t` always puts the executables under bin/. However, "real"
virtualenv will use Script/ instead of bin/ on Windows, even for
binaries.

We were dealing with this for virtualenv's activate script by
replacing `/bin/` with `/*/` and relying on no other subdirectories
also having a file matching the name of the executable we're looking
for. We make this a little more reliable by now also matching the
first letter of the directory: `/[bS]*/`.

We also fix several areas other than `/[bS]*/activate` where we
weren't handling a Script/ directory at all.

This also required a minor tweak to how the interpreter mismatch
warning is generated in order to give the actual path to the virtual
environment interpreter rather than the the literal `/[bS]*/` glob
pattern.