pactivate: Supress spurious warnings
On earlier versions of Python (e.g., 3.6 or 2.7), git-pip.py produces
messages like "WARNING: You are using pip version 18.1; however, version
21.3.1 is available." This is spurious; by the time get-pip.py completes
the latest version of Pip is installed. We add --disable-pip-version-check
to the get-pip.py options to suppress this.

When installing the bootstrap version of virtualenv, we get the message
"WARNING: Target directory …/.build/bootstrap/pactivate3.6/bin already
exists. Specify --upgrade to force replacement." Presumably this is because
`pip install -t …` expects the target directory to be empty, but we have
the pip module and program already installed in there. As suggested, we add
the --upgrade option to suppress this message.

We also, for the first time, need to modify the version number to indicate
an unreleased change, so document (in CHANGELOG.md) how we do that.