WIP pactivate: Use Windows `py` if available; add win-test
XXX CHANGELOG: date for 0.4.0

Prevously when not using a $PROJ/.python link to find the Python
interpreter we used `python3`, relying on the shell $PATH search to
find the executable with each invocation. Now we do a better search,
using the Windows Python Launcher `py` if we're in a MINGW environment
and can find it, otherwise finding the full path to the first
`python3` or `python` in $PATH.

We also add a new `win-test` script allowing us to do a basic test on
Windows. This covers minimal functionality and does not replace any of
the tests run by `Test`. It shares a lot of code with `cont-test`; the
common code should be extracted at some point.

pactivate details:
* The `.python` link to specify a specific interpreter doesn't work under
  Windows due to lack of proper support for symbolic links. This is
  documented and we're considering ways to work around this.
* Slightly more deferencing than before: $__pa_python is now always an
  absolute path to the interpreter, rather needing to be found via $PATH.
  $PATH. This allows us to do e.g., [[ -L $__pa_python ]]) on the filename,
  though we won't use this facility until later.
* Bootstrap virtualenv install does not use the --upgrade option on Windows
  because it aborts the install with "PermissionError: [WinError 5] Access
  is denied: '…\\bootstrap\\pactivate\\bin\\pip.exe'".