WIP Test: .no-sudo file disables sudo; MSYS support for Windows
XXX update README for .no-sudo

Creating a file `.no-sudo` in the project directory will instruct Test
not to use `sudo` to run `docker`. This is useful for (arguably poorly
configured) systems where the user has access to the Docker daemon's
socket (and thus effectly has root access on the system) but cannot
sudo to root.

If running in an MSYS environment on Windows we need to do several
things to make this work:
1. Don't use sudo; it's not available in this environment.
2. Run `docker` under `winpty` because `docker run` requires a Windows
   Console interface when given the -t option. 3.
3. The path of the program to run within the container is absolute and
   so must be prefixed with an extra `/` to keep MSYS from translating
   it into a Windows path (`C:/Program Files/Git/test/cont-test`).

It is possible to use a relative path to avoid the translation but
this has two issues. First, it relies on the CWD being set to `/` in
the container, and so becomes more fragile. Second, it's not obvious
that if someone later changes the path to an absolute path it will
break operation under Windows, and it's a reasonable assumption that
developers on Linux won't be testing every commit on Windows as well.

We did try to use MSYS_NO_PATHCONV=1 [1] and MSYS2_ARG_CONV_EXCL='*' [2]
but while both of these suppress path conversion with other programs
(e.g., `py` to run Python), they do not seem to do so for arguments
given to `docker`. It's not clear why this is.

[1]: https://stackoverflow.com/a/48348531/107294
[2]: https://www.msys2.org/docs/filesystem-paths/