Test: Do not use sudo if a `.no-sudo` file is found
This is useful for people on systems where they can't sudo. In
particular it's needed on Windows where there is no `sudo` command.
(The tests do not yet run on Windows, however, due to other issues
relating to using the `docker` command in MSYS Windows environments.)

Normally files that configure how the tests are run would not go in
the `.gitignore` file so that they will appear in `git status` output
to remind the developer that she's using a different configuration. In
this case, however, this does not change how the tests themselves run,
so we include it in `.gitignore` in order to help prevent accidentally
committing that file.