Test: Allow MSYS/MINGW Windows Bash users to run Docker tests
(This does NOT make pactivate work on Windows. That's coming later.)

If running in a MINGW/MSYS environment on Windows we need to do
several things to be able to run a Docker container:
1. Don't use sudo; it's not available in this environment.
2. Run `docker` under `winpty` because `docker run` requires a
   Windows Console interface when given the -t option.
3. The path of the program to run in the container, `/test/cont-test`,
   is absolute so we must prefix it with an extra `/` to keep MINGW
   from translating it into a Windows path such as
   `C:/Program Files/Git/test/cont-test`.

It is possible to use a relative path to avoid the translation, but
this has two issues. First, it relies on the CWD in the container
being set to `/` and so becomes more fragile. Second, with a relative
path there it's not obvious that if someone later changes the path to
an absolute path it will break operation under Windows, and it's a
reasonable assumption that developers on Linux won't be testing every
commit on Windows as well.

We did try to use MSYS_NO_PATHCONV=1 [1] and MSYS2_ARG_CONV_EXCL='*' [2]
but while both of these suppress path conversion with other programs
(e.g., `py` to run Python), they do not seem to do so for arguments
given to `docker`. It's not clear why this is.

[1]: https://stackoverflow.com/a/48348531/107294
[2]: https://www.msys2.org/docs/filesystem-paths/