pactivate: Bootstrap pip install now checks for and returns failure
The bootstrap pip install can fail on certain machine configurations, such
as a Debian system without the python3-distutils package installed. We
didn't check for these failures and so tried to carry on, eventually
cascading into other failures if `set -e` was not in effect (which it
usually isn't) and returning success instead of a failure exit code.

We now check this and return the failure code if it failed.