pactivate: Print better error message when distutils.cmd not present
The module `distutil.cmd` is required by the pip bootstrapper, and that
module is included in the standard Python distribution. Debian, however,
does not incude it in the `python3` package but in the `python3-distutils`
package, so it may not be present on a Debian-based system even when the
Python 3 interpreter is present.

The error message produced in such a case, a very long traceback followed
by "ModuleNotFoundError: No module named 'distutils.cmd'", is not very
informative about what the problem is. So we now first test to see if we
can import `distutils.cmd` and, if we cannot, fail with a suggestion that
the user may want to `apt-get install `python3-distutils`. That's exact
sugestion is only valid for Debian and its derivatives, but the user should
still get the general idea even if he's using another distribution that has
a similar issue.