pactivate: Fix prompt
The prompt when you activate a virtualenv is supposed to be just the
basedir name alone, without a full path prefix. But we were accidentally
using `dirname` instead of `basename` on the basedir, and so the prompt
became the path up to and not including the actual base directory itself.