pae -r: Better "command is virt env name" detection
As before, in run mode if no arguments are given after the virtual
env name we assume that the command name is the same as the virtual
env name.

However, if any arguments are given, we now check to see if the first
argument is an executable in the virtual env's bin/ directory. If it
is, we run that command, otherwise we assume that the command is the
same as the virtual env name.

This allows us to avoid typing the command name twice for virtual
environments dedicated to a particular command, i.e., we can use
`pae -r grip --version` instead of `pae -r grip grip --version`.