pae: exit with an error if the file is run instead of sourced, or not Bash
One shouldn't be able to run this directly, since it's not executable, but
in case someone does mark it so and try to run it, or uses `bash pae`, we
exit with an error message rather than create all the functions and then
exit the child process, leaving the user perhaps to wonder why the script
does nothing.

We also exit with an error if $BASH_VERSION is not set, since we use
a fair number of bash-isms in the functions.