pae: Add command line completion
This always does completion for the run action, regardless of
what the action really is. The other actions shouldn't try to
complete an excecutable file under the virtual env bin directory,
but this will do for now.

This required significant debugging using the new __pae_debug()
function; the debugging statements, though annoying, have been
left in (but disabled) because it's really hard to work on this
without those.