tscript/docker-data/pkg-inst: Extract install script from Dockerfile
This will let us get more sophisticated with our installs, so we can
detect which package system we need to use based on the image name.

Note that we install all other scripts _after_ we run the time-consuming
pkg-inst so that we don't need to re-run it when changing those other
scripts.