pkg-inst: Install bash, curl when using official Python Alpine images
This is enough to let us start running /test/cont-test, but that won't pass
yet because it's doing a bunch of Debian-specific stuff related to the
distutils tests. We need to figure out how to deal with that and the
python2 tests that need be run only one Debian and Ubuntu systems, and
possibly only on certain versions of those.