cont-test: Disable some tests when not on Debian-derived distros
The test covering altbase (-b and -B) and requirements.txt is from the
Debian Python 2 test so that these options are tested on all Python 3
systems.

This still leaves us with the (manual) 'checking .python difference
warning' test executed only if the Python 2 test is executed since that's
the only case right now where we use a second, different Python interpreter
in a single container. It's not that important of a test, but it would be
nice to figure out a way to do this better.

Everything seems now to be running just fine in a python:3.10-alpine
container, so add that to the standard list of containers to test.