cont-test: Proper test for .python interpreter difference warning
We now check this in the main tests, rather than the special Debian-only
tests. What was holding us back from doing this before was the lack of a
second, different Python interpreter binary, but we fix this by just making
a copy of the standard binary and appending a newline to it. This runs just
fine, but is now technically a different file so that the compare with the
existing binary will fail.