tscript/docker-data/bash_profile: Source .bashrc for Test -i option
When tscript/pactivate-test-docker is run with the -i option it starts a
login shell with `bash -l`. This was not sourcing our .bashrc because only
interactive shells that are _not_ login shells do this. So we add a
.bash_profile that sources our .bashrc to make this work.

We do not need to add -i to the `bash -l` command that starts the
interactive shell because it will be connected to a tty (unless the user
has redirected stdin) and so will be automatically interactive anyway.
(If stdin is redirected bash will fail, probably because of the -l option,
but non-interactive shells with -i don't seem worth supporting.)