pactivate 0.3.6: Fix spurious bootstrap pip re-install
On the first run of pactivate after an initial `pactivate` bootstrap
it was downloading and installing the bootstrap pip again, though it
was no longer needed. This was due to some interesting behaviour of
the --update option.

The fix for this is simple: restore bin/pip by requesting an "upgrade"
of the bootstrap pip when we install the bootstrap virtualenv.