tscript/pactivate.arg: Default arguments extracted from Test
Move the sequence of default arguments with which tscript/pactivate is run
into a stand-alone file. This makes it easier to see what they are and this
file will also be the default list for `git tscript` when that appears.

To allow for comments, we remove any lines starting with `#`. Blank lines
will also be ignored as they're simply whitespace in the shell expansion.
(This means we can't have args with spaces in them.)