tscript/pactivate: Fix tests to work with Python ≤3.6
The tests previously assumed that the bootstrap directory was always
named $build/bootstrap/pactivate, and thus couldn't handle older versions
of Python (3.5 or earlier) where we use an alternate version of get-pip.py
(and change the directory name so that we don't accidentally use the
alternate version for a "modern" version of Python).

We now dynamically determine the actual bootstrap directory name after
the initial run of pactivate.

This allows us to enable the tests for versions 3.5 and 3.6. We cannot
test 3.0 through 3.4 because the yaml package we use to test
requirements.txt doesn't work on those versions, and we cannot (yet)
test 2.7 because the interpreter is not named `python3`.