pactivate: Use `python` if `python3 --version` fails
This now allows our tests to pass on the python:2.7-alpine image as well
as (presumably) other systems that provide `python` but not `python3`.

This should also now correctly use `py` on Windows/MINGW systems, but that
has not been tested yet.

tscript/pactivate now also indicates exactly which Python interpreter
(including where it's found in the path) and version it's being run
against.

This bumps our pactivate version from 0.3.x to 0.4.0.