cont-test: Replace yaml with six as requirements.txt test package
Six is a single pure Python file, and one of the simplest PyPI
packages available, unlike pyyaml which uses C code. This allows us
now to test Python 3.4, which we have enabled.

Also, it's not yet seen here, but pyyaml also doesn't work with the
Docker python:2.7-alpine image due to some sort of compiliation
problems, whereas six should.

(We've not changed the test with Python 2.7 on debian:11, where pyyaml
does work. We should consider how we might reduce the scope of those
tests in general, given the additional coverage we've added over the
last little while.)