pactivate: Do not test for distutils on Python ≥ 3.11
This removes a deprecation warning for 3.11, and makes it work for 3.12.

distutils.cmd is necessary for pactivate to work properly in Pythons ≤3.10,
and we check for it so that we can print a good error message rather than
just surfacing a mysterious ModuleNotFoundError.

But distutils is both unneeded and deprecated in 3.11, and removed in 3.12,
so we don't check for it in versions from 3.11 upward.