pactivate: Increase curl timeout and add retries
For some reason the test on python:3.11-alpine was failing with:

    curl: (7) Failed to connect to bootstrap.pypa.io port 443 \
        after 2001 ms: Couldn't connect to server

That happens only on this one image, and it's an awfully short timeout,
so perhaps it's something in the configuration of that image.

At any rate, we now explictly specfy a 10-second connect timeout (this
is just the timeout for the initial connection to complete, not the
whole transfer) and retry 3 times instead of the default of 0 retries.