pactivate 0.5.1: IPv4 retry fix for curl timeout issue
The curl timeout issues we'd thought were fixed in 0.5.0 re-appeared;
debugging with `curl -v` indicates that for some reason on the
python:3.9-alpine (which was where we had the problem last time) it
was sometimes resolving only an IPv6 address for bootstrap.pypa.io,
even though there is also an IPv4 address available.

We work around this issue by, on a failure to fetch, retrying with
`curl --ipv4` (to restrict it to IPv4 only); this seems to fix the
issue.

The previous timeout arguments are now removed; the default timeouts
should be fine.