pae: Fix mysterious issue with adjacent slashes in path when running python
I can't seem to replicate this in the automated tests, but I have
situations where running `pae ./ python` or `pae .//somedir python`
failed due to multiple adjacent slashes in the path to `python`.
(The former produces two adjacent slashes between the directory name
and the subdirs we append to it when we do `$vedir/...`.)

In this cases python would fail to properly add the virtual environment's
libraries to sys.path, and so importing something from the virtual
environment would fail.

This has been observed to fix the problem in my environment, and also (via
a debugging printf) to properly remove the multiple slashes when running in
tscript/pae.