pactivate: Use `python -m pip` instead of `pip` to work under Windows
This makes no difference either way for Linux, but under Windows we were
running a pip.exe that was tryiing to replace itself and failing because
that pip.exe file was in use "by another process" (that being the pip
instance that was trying to do the upgrade).

This completes initial support for running under MINGW on Windows.