pactivate: (v0.5.4) Upgrade pip during virtualenv install
Often get_pip.py is a version of pip old enough that it will generate
messages suggesting you upgrade to a new version of pip, which can
appear again and again in build scripts running pip in the virtual
environment. We now automatically upgrade pip to the latest version as
part of the virtualenv setup process so that we avoid this.