tscript/pae: Note that these tests use the production version of pactivate
This isn't ideal, but it's going to take some thought to see how we can
change them to use the test version while still testing download and so on.
(Perhaps we can do all the download tests first, and then replace the
downloaded version with our version? Or perhaps we should be able to
specify the download URL.)