pactivate: Fix virtualenv pip update to work on Windows
We must use `python -m pip` instead of `pip` to upgrade pip itself
because pip cannot replace itself on Windows when it's running due to
the file being locked while running.

We knew this from the same problem earlier on in the bootstrap
process, but hadn't applied here with the new code.