pactivate 0.5.7: Allow use with ksh and zsh
This was not only complaining if you weren't using Bash, but if you
sourced the file it would make the calling shell exit. (Ksh/zsh users
often saw this as just their shell window mysteriously closing.)

We now do the "is it being sourced" check (which works only in Bash) first,
then checks we're using one of Bash, ksh and zsh, and after that print a
warning if Bash isn't being used. (This helps with debugging ksh and zsh
issues, since both are little-tested.)

I see no obvious way to make the "is it being sourced" check work in ksh or
zsh and, apparently, neither do the virtualenv developers who also have a
check in their `activate` script that works only on Bash; for ksh and zsh
users who run instead of sourcing the script it just carries on, leaving
them (perhaps mysteriously) without an activated virtual environment.

Zsh does already have a non-critical issue: it produces the following
warning, which Bash and ksh don't:

    WARNING: …/.build/virtualenv/bin/python (…) is not python3 (…)

This warning is spurious; investigation shows that the `python`
in the virtualenv is a link to /usr/bin/python3. Thus our test
for this is not working properly under Zsh.