tscript/Dockerfile/demopkg/: local package to test install via relative path
Installs via relative paths in `requirements.txt` were previously not
working; this was fixed in the last commit. This adds a test for it;
the entire package install should fail before even the test for the
presence of `six` if the package is not being looked for in the
project directory instead of the current working directory. The commit
is separate so that we can test that it fails against the commit
before the fix, though this can also be tested by placing the
following after the lines that set up $altbase/requirements.txt:

    mv /test/test-requirements/demopkg ./

`Dockerfile` puts the local package in the expected place, which is
maybe a bit mysterious when you read the actual test code that expects
it to be there magically. But this is easier than trying to keep track
of various copies of the package dir and copy them around.

We considered using `-e ./demopkg/`, but that doesn't work in Python
2.7; apparently in that version pyproject.toml works for non-editable
installs but we need to use a setup.py for an editable install.